feat(net): SSE-over-TLS in RemoteMetaSync + insecureSkipVerify UI (#69)
Extract the TLS/socket plumbing into a header-only HttpClientTls.h so
RemoteMetaSync's SSE long-poll can share it. Previously /meta over
https:// silently degraded to short-polling because runSSE() used raw
sockets; now it goes through the same TLS-aware Connection HttpClient
uses, so HTTPS hosts get a real event stream.

Adds an Advanced toggle "Allow self-signed TLS cert (dev only)" under
Streaming → Public Directory. Flag is plumbed through DirectoryClient::
Config to every register/heartbeat/patch/delete call and persisted in
the config file. Off by default — only for self-signed dev hosts.

Drive-by: deletes the dead `httpGet` helper in RemoteMetaSync (long
since superseded by httpGetViaAvio).

Header uses `static constexpr` (not `inline constexpr`) for INVALID_SOCK
so MXE's older MinGW-w64 GCC accepts it.
