+++ src/streaming/DirectoryClient.cpp
@@ -285,9 +285,12 @@ bool DirectoryClient::doRegister()
         {"version",          cfg.version},
     };
 
+    HttpClient::Options opts;
+    opts.insecureSkipVerify = cfg.insecureSkipVerify;
     auto resp = HttpClient::send(HttpClient::Method::POST,
                                  cfg.directoryUrl + "/register",
-                                 body.dump());
+                                 body.dump(),
+                                 opts);
     if (!resp.ok)
     {
         setError("register transport failure: " + resp.error);
@@ -329,13 +332,15 @@ bool DirectoryClient::doRegister()
 bool DirectoryClient::doHeartbeat()
 {
     std::string baseUrl, streamId, ownerToken;
+    bool        insecureSkipVerify;
     int cc;
     {
         std::lock_guard<std::mutex> lock(m_mu);
-        baseUrl    = m_cfg.directoryUrl;
-        streamId   = m_streamId;
-        ownerToken = m_ownerToken;
-        cc         = m_currentClientCount;
+        baseUrl            = m_cfg.directoryUrl;
+        streamId           = m_streamId;
+        ownerToken         = m_ownerToken;
+        insecureSkipVerify = m_cfg.insecureSkipVerify;
+        cc                 = m_currentClientCount;
     }
     if (streamId.empty() || ownerToken.empty()) return false;
 
@@ -344,9 +349,12 @@ bool DirectoryClient::doHeartbeat()
         {"ownerToken",  ownerToken},
         {"clientCount", cc},
     };
+    HttpClient::Options opts;
+    opts.insecureSkipVerify = insecureSkipVerify;
     auto resp = HttpClient::send(HttpClient::Method::POST,
                                  baseUrl + "/heartbeat",
-                                 body.dump());
+                                 body.dump(),
+                                 opts);
     if (!resp.ok)
     {
         setError("heartbeat transport: " + resp.error);
@@ -387,15 +395,17 @@ bool DirectoryClient::doPatchIfPending()
 {
     Patch p;
     std::string baseUrl, streamId, ownerToken;
+    bool        insecureSkipVerify;
     {
         std::lock_guard<std::mutex> lock(m_mu);
         if (!m_hasPendingPatch) return true;
         p = m_pendingPatch;
         m_pendingPatch = Patch{};
         m_hasPendingPatch = false;
-        baseUrl    = m_cfg.directoryUrl;
-        streamId   = m_streamId;
-        ownerToken = m_ownerToken;
+        baseUrl            = m_cfg.directoryUrl;
+        streamId           = m_streamId;
+        ownerToken         = m_ownerToken;
+        insecureSkipVerify = m_cfg.insecureSkipVerify;
     }
     if (streamId.empty() || ownerToken.empty()) return false;
 
@@ -411,9 +421,12 @@ bool DirectoryClient::doPatchIfPending()
     if (!p.endpoint.empty())     body["endpoint"]     = p.endpoint;
     if (!p.endpointMode.empty()) body["endpointMode"] = p.endpointMode;
 
+    HttpClient::Options opts;
+    opts.insecureSkipVerify = insecureSkipVerify;
     auto resp = HttpClient::send(HttpClient::Method::PATCH,
                                  baseUrl + "/streams/" + streamId,
-                                 body.dump());
+                                 body.dump(),
+                                 opts);
     if (!resp.ok)
     {
         setError("patch transport: " + resp.error);
@@ -434,11 +447,13 @@ bool DirectoryClient::doPatchIfPending()
 bool DirectoryClient::doDelete()
 {
     std::string baseUrl, streamId, ownerToken;
+    bool        insecureSkipVerify;
     {
         std::lock_guard<std::mutex> lock(m_mu);
-        baseUrl    = m_cfg.directoryUrl;
-        streamId   = m_streamId;
-        ownerToken = m_ownerToken;
+        baseUrl            = m_cfg.directoryUrl;
+        streamId           = m_streamId;
+        ownerToken         = m_ownerToken;
+        insecureSkipVerify = m_cfg.insecureSkipVerify;
     }
     if (streamId.empty() || ownerToken.empty())
     {
@@ -447,9 +462,12 @@ bool DirectoryClient::doDelete()
     }
 
     json body = {{"ownerToken", ownerToken}};
+    HttpClient::Options opts;
+    opts.insecureSkipVerify = insecureSkipVerify;
     auto resp = HttpClient::send(HttpClient::Method::DELETE_,
                                  baseUrl + "/streams/" + streamId,
-                                 body.dump());
+                                 body.dump(),
+                                 opts);
     if (!resp.ok)
     {
         LOG_WARN("DirectoryClient::doDelete — transport: " + resp.error);
